Example related fixes.   - Use floats for timeouts.   - Add BaseService in favor of ServerServiceStub.

+++ quarkc/test/emit/expected/java/builtin/src/main/java/builtin/Service.java
@@ -4,6 +4,6 @@ public interface Service {
     public static builtin.reflect.Class builtin_Service_ref = builtin_md.Root.builtin_Service_md;
      String getName();
      ServiceInstance getInstance();
-     Long getTimeout();
+     Double getTimeout();
      builtin.concurrent.Future rpc(String methodName, java.util.ArrayList<Object> args);
 }
